⚡️ :: (#90) ErrorResponse 대신 BaseResponse 사용
